moving decoder into its own module
