+++ src/api/mod.rs
@@ -1,7 +1,4 @@
 pub mod evaluator;
 pub use evaluator::Evaluator;
-pub(crate) mod decoder;
 pub mod external_reader;
 pub mod reader;
-
-pub use decoder::pkl_eval_module;
