+++ src/pkl/deserializer.rs
@@ -301,7 +301,7 @@ mod tests {
             password: String,
         }
 
-        let ast = Value::Array(vec![
+        let eval_module_response = Value::Array(vec![
             Value::Integer(1.into()),
             Value::String("example".into()),
             Value::String("file:///Users/testing/rpkl/examples/example.pkl".into()),
@@ -352,8 +352,8 @@ mod tests {
             ]),
         ]);
 
-        let pkl_mod =
-            crate::api::decoder::pkl_eval_module(&ast).expect("failed to evaluate pkl ast");
+        let pkl_mod = crate::decoder::decode_module(&eval_module_response)
+            .expect("failed to evaluate pkl ast");
         let mapped = pkl_mod
             .serialize_pkl_ast()
             .expect("failed to serialize pkl module");
@@ -370,7 +370,99 @@ mod tests {
                 password: "secret".to_owned(),
             },
         };
-        assert_eq!(expected, deserialized)
+        assert_eq!(expected, deserialized);
+    }
+
+    #[test]
+    fn deserialize_time() {
+        #[cfg(feature = "dhat-heap")]
+        let _profiler = dhat::Profiler::new_heap();
+
+        #[derive(Debug, PartialEq, Deserialize)]
+        struct Config {
+            ip: String,
+            port: u16,
+            birds: Vec<String>,
+            database: Database,
+        }
+
+        #[derive(Debug, PartialEq, Deserialize)]
+        struct Database {
+            username: String,
+            password: String,
+        }
+
+        let ast = Value::Array(vec![
+            Value::Integer(1.into()),
+            Value::String("example".into()),
+            Value::String("file:///Users/testing/code/rust/rpkl/examples/example.pkl".into()),
+            Value::Array(vec![
+                Value::Array(vec![
+                    Value::Integer(16.into()),
+                    Value::String("ip".into()),
+                    Value::String("127.0.0.1".into()),
+                ]),
+                Value::Array(vec![
+                    Value::Integer(16.into()),
+                    Value::String("port".into()),
+                    Value::Integer(8080.into()),
+                ]),
+                Value::Array(vec![
+                    Value::Integer(16.into()),
+                    Value::String("birds".into()),
+                    Value::Array(vec![
+                        Value::Integer(5.into()),
+                        Value::Array(vec![
+                            Value::String("Pigeon".into()),
+                            Value::String("Hawk".into()),
+                            Value::String("Penguin".into()),
+                        ]),
+                    ]),
+                ]),
+                Value::Array(vec![
+                    Value::Integer(16.into()),
+                    Value::String("database".into()),
+                    Value::Array(vec![
+                        Value::Integer(1.into()),
+                        Value::String("Dynamic".into()),
+                        Value::String("pkl:base".into()),
+                        Value::Array(vec![
+                            Value::Array(vec![
+                                Value::Integer(16.into()),
+                                Value::String("username".into()),
+                                Value::String("admin".into()),
+                            ]),
+                            Value::Array(vec![
+                                Value::Integer(16.into()),
+                                Value::String("password".into()),
+                                Value::String("secret".into()),
+                            ]),
+                        ]),
+                    ]),
+                ]),
+            ]),
+        ]);
+        let expected = Config {
+            ip: "127.0.0.1".into(),
+            port: 8080,
+            birds: vec!["Pigeon".into(), "Hawk".into(), "Penguin".into()],
+            database: Database {
+                username: "admin".to_owned(),
+                password: "secret".to_owned(),
+            },
+        };
+        // let now = std::time::Instant::now();
+        let pkl_mod = crate::decoder::decode_module(&ast).expect("failed to evaluate pkl ast");
+        let mapped = pkl_mod
+            .serialize_pkl_ast()
+            .expect("failed to serialize pkl module");
+
+        let deserialized = Config::deserialize(&mut Deserializer::from_pkl_map(&mapped))
+            .expect("failed to deserialize");
+
+        // let elapsed = now.elapsed();
+        // print_time!(elapsed);
+        assert_eq!(expected, deserialized);
     }
 
     #[test]
